implemented Api Getting in Unity

+++ RobotWebApi/RobotWebApi/Models/Camera.cs
@@ -2,8 +2,8 @@
 {
     public class Camera
     {
-        public int Width { get; set; }
-        public int Height { get; set; }
+        public float Width { get; set; }
+        public float Height { get; set; }
 
         public override string ToString()
         {
